TimeSeriesChart refactor to remove unnecessary container component
