+++ ui/panels-plugin/src/plugins/time-series-chart/TimeSeriesChartPanel.tsx
@@ -11,35 +11,132 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { PanelProps, useTimeSeriesQueries } from '@perses-dev/plugin-system';
+import { PanelProps, useTimeSeriesQueries, useTimeRange } from '@perses-dev/plugin-system';
+import { useMemo } from 'react';
+import { GridComponentOption } from 'echarts';
+import { Box, Skeleton } from '@mui/material';
+import { LineChart, EChartsDataFormat, ZoomEventData } from '@perses-dev/components';
 import { useSuggestedStepMs } from '../../model/time';
+import { StepOptions, ThresholdColors, ThresholdColorsPalette } from '../../model/thresholds';
 import { TimeSeriesChartOptions } from './time-series-chart-model';
-import { TimeSeriesChartContainer } from './TimeSeriesChartContainer';
+import { getLineSeries, getCommonTimeScale, getYValues, getXValues } from './utils/data-transform';
+
+export const EMPTY_GRAPH_DATA = {
+  timeSeries: [],
+  xAxis: [],
+};
 
 export type TimeSeriesChartProps = PanelProps<TimeSeriesChartOptions>;
 
 export function TimeSeriesChartPanel(props: TimeSeriesChartProps) {
   const {
-    spec: { queries, show_legend, thresholds, unit },
+    spec: { queries, show_legend, thresholds },
     contentDimensions,
   } = props;
 
+  const unit = props.spec.unit ?? {
+    kind: 'Decimal',
+    decimal_places: 2,
+  };
+
   const suggestedStepMs = useSuggestedStepMs(contentDimensions?.width);
   const queryResults = useTimeSeriesQueries(queries, { suggestedStepMs });
+
+  const { setTimeRange } = useTimeRange();
+
+  // populate series data based on query results
+  const { graphData, loading } = useMemo(() => {
+    const timeScale = getCommonTimeScale(queryResults);
+    if (timeScale === undefined) {
+      for (const query of queryResults) {
+        // does not show error message if any query is successful (due to timeScale check)
+        if (query.error) throw query.error;
+      }
+      return {
+        graphData: EMPTY_GRAPH_DATA,
+        loading: true,
+      };
+    }
+
+    const graphData: EChartsDataFormat = { timeSeries: [], xAxis: [], rangeMs: timeScale.endMs - timeScale.startMs };
+    const xAxisData = [...getXValues(timeScale)];
+
+    let queriesFinished = 0;
+    for (const query of queryResults) {
+      // Skip queries that are still loading and don't have data
+      if (query.isLoading || query.data === undefined) continue;
+
+      for (const timeSeries of query.data.series) {
+        const yValues = getYValues(timeSeries, timeScale);
+        const lineSeries = getLineSeries(timeSeries.name, yValues);
+        graphData.timeSeries.push(lineSeries);
+      }
+      queriesFinished++;
+    }
+    graphData.xAxis = xAxisData;
+
+    if (thresholds !== undefined && thresholds.steps !== undefined) {
+      const defaultThresholdColor = thresholds.default_color ?? ThresholdColors.RED;
+      thresholds.steps.forEach((step: StepOptions, index: number) => {
+        const stepPaletteColor = ThresholdColorsPalette[index] ?? defaultThresholdColor;
+        const thresholdLineColor = step.color ?? stepPaletteColor;
+        const stepOption: StepOptions = {
+          color: thresholdLineColor,
+          value: step.value,
+        };
+        const thresholdName = step.name ?? `Threshold ${index + 1} `;
+        const thresholdData = Array(xAxisData.length).fill(step.value);
+        const thresholdLineSeries = getLineSeries(thresholdName, thresholdData, stepOption);
+        graphData.timeSeries.push(thresholdLineSeries);
+      });
+    }
+
+    return {
+      graphData,
+      loading: queriesFinished === 0,
+      allQueriesLoaded: queriesFinished === queryResults.length,
+    };
+  }, [queryResults, thresholds]);
 
   if (contentDimensions === undefined) {
     return null;
   }
 
-  // TODO: Do we need this container component any more now that we can run multiple queries here?
+  if (loading === true) {
+    return (
+      <Box
+        sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center' }}
+        width={contentDimensions.width}
+        height={contentDimensions.height}
+      >
+        <Skeleton variant="text" width={contentDimensions.width - 20} height={contentDimensions.height / 2} />
+      </Box>
+    );
+  }
+
+  const legendOverrides = {
+    show: show_legend === true,
+    type: 'scroll',
+    bottom: 0,
+  };
+
+  const gridOverrides: GridComponentOption = {
+    bottom: show_legend === true ? 35 : 0,
+  };
+
+  const handleDataZoom = (event: ZoomEventData) => {
+    // TODO: add ECharts transition animation on zoom
+    setTimeRange({ start: new Date(event.start), end: new Date(event.end) });
+  };
+
   return (
-    <TimeSeriesChartContainer
-      width={contentDimensions.width}
+    <LineChart
       height={contentDimensions.height}
+      data={graphData}
       unit={unit}
-      show_legend={show_legend}
-      thresholds={thresholds}
-      queryResults={queryResults}
+      legend={legendOverrides}
+      grid={gridOverrides}
+      onDataZoom={handleDataZoom}
     />
   );
 }
